+++ src/wildlint/__init__.py
@@ -18,7 +18,7 @@ from .property_templates import (
     get_template,
 )
 
-__version__ = "0.8.6"
+__version__ = "0.8.7"
 
 __all__ = [
     "CHECKERS",